requirements.txt
Flask==3.0.3
psycopg2-binary
pymongo==4.8.0
python-jose==3.3.0
python-dateutil==2.8.2
python-dotenv==1.0.1
flask-cors==5.0.0
pyyaml==6.0.1
